# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "My Brand API",
        "description": "This is My Brand API",
        "version": "1.0.0",
        "contact": {
            "email": "niyonagboaristide17@gmail.com"
        }
    },
...
    "servers": [
        {
            "url": "http://localhost:3000",
            "name": "DEV"
        }
    ],
    "paths": {
...
        "/api/v1/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get One Article",
                "description": "Get  Article By Id",
                "operationId": "getArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artticle Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete Article",
                "description": "Delete One  Article",
                "operationId": "delArticle",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artticle Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }, "patch": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update Article",
                "description": "Update An Article",
                "operationId": "patchArticle",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true
                    },{
                        "name": "id",
                        "in": "path",
                        "description": "Artticle Id",
                        "required": true
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "content"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "type": "file",
                        "description": "image"
                    }
                    
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
...
}